Split hyperbolic geometry into geometry and graphical bits
